Refatorando para apenas um DadosDespesaCusteio
